Refactored to useSmartContractCall
