--- frontend/src/enums/JobVariables.ts
@@ -1,14 +1,0 @@
-export enum JobVariables {
-  MINIMUM_BOUNTY,
-  MINIMUM_DEPOSIT,
-  DAO_FEE,
-  RESOLUTION_FEE_PERCENTAGE,
-  MINIMUM_SPLIT_CHUNK_PERCENTAGE,
-  COMPLETED_TIMEOUT_SECONDS,
-  REPORT_DEPOSIT,
-  REPORT_TOKEN,
-  REPORT_REWARD_PERCENT,
-  daoTreasury,
-  disputeResolver,
-  jobCount,
-}
